+++ src/collective/collectionfilter/portlets/sorting.py
@@ -1,17 +1,20 @@
 # -*- coding: utf-8 -*-
-from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from collective.collectionfilter import _
 from collective.collectionfilter.baseviews import BaseSortOnView
-from collective.collectionfilter.interfaces import ICollectionFilterResultListSort  # noqa
+from collective.collectionfilter.interfaces import (  # noqa
+    ICollectionFilterResultListSort,
+)
 from collective.collectionfilter.portlets import BasePortletRenderer
 from plone.app.portlets.portlets import base
 from plone.portlets.interfaces import IPortletDataProvider
+from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from zope.interface import implementer
 
 
-class ICollectionFilterSortOnPortlet(ICollectionFilterResultListSort, IPortletDataProvider):  # noqa
-    """Portlet interface based on ICollectionFilterSchema
-    """
+class ICollectionFilterSortOnPortlet(
+    ICollectionFilterResultListSort, IPortletDataProvider
+):  # noqa
+    """Portlet interface based on ICollectionFilterSchema"""
 
 
 @implementer(ICollectionFilterSortOnPortlet)
@@ -21,17 +24,17 @@ class Assignment(base.Assignment):
     target_collection = None
     view_name = None
     sort_on = u""
-    input_type = 'links'
-    content_selector = '#content-core'
+    input_type = "links"
+    content_selector = "#content-core"
 
     def __init__(
         self,
         header=u"",
         target_collection=None,
         sort_on=u"",
-        input_type='links',
+        input_type="links",
         view_name=None,
-        content_selector='#content-core',
+        content_selector="#content-core",
     ):
         self.header = header
         self.target_collection = target_collection
@@ -42,8 +45,7 @@ class Assignment(base.Assignment):
 
     @property
     def portlet_id(self):
-        """Return the portlet assignment's unique object id.
-        """
+        """Return the portlet assignment's unique object id."""
         return id(self)
 
     @property
@@ -51,20 +53,18 @@ class Assignment(base.Assignment):
         if self.header:
             return self.header
         else:
-            return _(u'Collection Result Sorting')
+            return _(u"Collection Result Sorting")
 
 
 class Renderer(BasePortletRenderer, BaseSortOnView):
-    render = ViewPageTemplateFile('sorting.pt')
+    render = ViewPageTemplateFile("sorting.pt")
 
 
 class AddForm(base.AddForm):
 
     schema = ICollectionFilterSortOnPortlet
     label = _(u"Add Collection Result Listing Sort Portlet")
-    description = _(
-        u"This portlet shows sorting options for the result listing."
-    )
+    description = _(u"This portlet shows sorting options for the result listing.")
 
     def create(self, data):
         return Assignment(**data)
@@ -74,6 +74,4 @@ class EditForm(base.EditForm):
 
     schema = ICollectionFilterSortOnPortlet
     label = _(u"Edit Collection Result Listing Sort Portlet")
-    description = _(
-        u"This portlet shows sorting options for the result listing."
-    )
+    description = _(u"This portlet shows sorting options for the result listing.")